Add files via upload
PREDICTIVE MAINTENANCE OF ENGINES  - this particular project is done by Pre-Processing the data, Analyzing the pre-processed data, training the machine with pre-processed data using an appropriate machine learning algorithm, saving the model and its dependencies, and lastly building a Web application using flask that integrates with the built model. This approach based on past data is the one we used to build this project.